backend/requirements.txt
fastapi
uvicorn
pymongo
numpy
pandas
scikit-learn
joblib
python-multipart
xgboost
scipy
paho-mqtt
tensorflow